complaint_system/requirements.txt
Django==5.2.7
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.1
Pillow==12.0.0

